endpoints: default to rendezvous hashing
Does what it says on the tin.
